api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pydantic==2.11.4
google-genai==1.14.0
python-dotenv==1.2.2
pypdf==6.0.0
sqlalchemy==2.0.40
alembic==1.15.2